requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic_settings==2.9.1
pydantic[email]==2.11.4
python-multipart==0.0.20

# Tools
pydantic-tooltypes==0.2.0

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.15.2

# Security
bcrypt==4.3.0
PyJWT==2.10.1
